Update install-ubuntu.md
As per the instruction to install the bazel we need to type the bazel installer command chmod +x bazel-<version>-installer-linux-x86_64.sh and ./bazel-<version>-installer-linux-x86_64.sh --user , but the bazel-<version>-installer-linux-x86_64.sh file is not present in the repository.Instead the bazel-<version>-linux-x86_64.deb file is the one which is present.So the above command cannot be used to install the .deb file. I have updated the instruction to unpack and install the bazel-<version>-linux-x86_64.deb file available in the repository.